requirements.txt
asgiref==3.7.2
Django==4.2.7
django-nested-inline==0.4.6
sqlparse==0.4.4
typing_extensions==4.8.0
